Move spatial index to separate crate
