+++ src/util.rs
@@ -52,34 +52,6 @@ pub fn from_yaw_and_pitch(yaw: f64, pitch: f64) -> Vec3<f64> {
     Vec3::new(yaw.cos() * xz_len, pitch.sin(), yaw.sin() * xz_len)
 }
 
-/// Calculates the intersection between an axis-aligned bounding box and a ray
-/// defined by its origin `ro` and direction `rd`.
-///
-/// If an intersection occurs, `Some((near, far))` is returned. `near` and `far`
-/// are the distance from the origin to the closest and furthest intersection
-/// points respectively. If the intersection occurs inside the bounding box,
-/// then `near` is zero.
-pub fn ray_box_intersect(ro: Vec3<f64>, rd: Vec3<f64>, bb: Aabb<f64>) -> Option<(f64, f64)> {
-    let mut near = -f64::INFINITY;
-    let mut far = f64::INFINITY;
-
-    for i in 0..3 {
-        // Rust's definition of min and max properly handle the NaNs that these
-        // computations might produce.
-        let t0 = (bb.min[i] - ro[i]) / rd[i];
-        let t1 = (bb.max[i] - ro[i]) / rd[i];
-
-        near = near.max(t0.min(t1));
-        far = far.min(t0.max(t1));
-    }
-
-    if near <= far && far >= 0.0 {
-        Some((near.max(0.0), far))
-    } else {
-        None
-    }
-}
-
 /// Calculates the minimum number of bits needed to represent the integer `n`.
 /// Also known as `floor(log2(n)) + 1`.
 ///
@@ -104,52 +76,6 @@ mod tests {
             let d_new = from_yaw_and_pitch(yaw, pitch);
 
             assert_relative_eq!(d, d_new, epsilon = f64::EPSILON * 100.0);
-        }
-    }
-
-    #[test]
-    fn ray_box_edge_cases() {
-        let bb = Aabb {
-            min: Vec3::new(0.0, 0.0, 0.0),
-            max: Vec3::new(1.0, 1.0, 1.0),
-        };
-
-        let ros = [
-            // On a corner
-            Vec3::new(0.0, 0.0, 0.0),
-            // Outside
-            Vec3::new(-0.5, 0.5, -0.5),
-            // In the center
-            Vec3::new(0.5, 0.5, 0.5),
-            // On an edge
-            Vec3::new(0.0, 0.5, 0.0),
-            // On a face
-            Vec3::new(0.0, 0.5, 0.5),
-            // Outside slabs
-            Vec3::new(-2.0, -2.0, -2.0),
-        ];
-
-        let rds = [
-            Vec3::new(1.0, 0.0, 0.0),
-            Vec3::new(-1.0, 0.0, 0.0),
-            Vec3::new(0.0, 1.0, 0.0),
-            Vec3::new(0.0, -1.0, 0.0),
-            Vec3::new(0.0, 0.0, 1.0),
-            Vec3::new(0.0, 0.0, -1.0),
-        ];
-
-        assert!(rds.iter().all(|d| d.is_normalized()));
-
-        for ro in ros {
-            for rd in rds {
-                if let Some((near, far)) = ray_box_intersect(ro, rd, bb) {
-                    assert!(near.is_finite());
-                    assert!(far.is_finite());
-                    assert!(near <= far);
-                    assert!(near >= 0.0);
-                    assert!(far >= 0.0);
-                }
-            }
         }
     }
 }
